move desktop handling into desktop{,_test}.go
