+++ snappy/click.go
@@ -20,10 +20,8 @@
 package snappy
 
 import (
-	"bufio"
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -478,129 +476,6 @@ func addPackageBinaries(m *snapYaml, baseDir string) error {
 func removePackageBinaries(m *snapYaml, baseDir string) error {
 	for _, app := range m.Apps {
 		os.Remove(generateBinaryName(m, app))
-	}
-
-	return nil
-}
-
-// this is a bit confusing, a localestring in xdg is:
-// "key=value" or "key[locale]=" with locale as:
-//  lang_COUNTRY@MODFIER (or a subset of this)
-var desktopFileI18nPattern = `(|\[[a-zA-Z_@]+\])`
-var validDesktopFileLines = []*regexp.Regexp{
-	// headers
-	regexp.MustCompile(`^\[Desktop Entry\]$`),
-	regexp.MustCompile(`^\[Desktop Action`),
-	// whitespace lines
-	regexp.MustCompile(`^\s*$`),
-	// lines with comments
-	regexp.MustCompile(`^\s*#`),
-	// https://specifications.freedesktop.org/desktop-entry-spec/latest/ar01s05.html
-	regexp.MustCompile(`^Type=`),
-	regexp.MustCompile(`^Version=`),
-	regexp.MustCompile(fmt.Sprintf(`^Name%s=`, desktopFileI18nPattern)),
-	regexp.MustCompile(fmt.Sprintf(`^GenericName%s=`, desktopFileI18nPattern)),
-	regexp.MustCompile(`^NoDisplay=`),
-	regexp.MustCompile(fmt.Sprintf(`^Comment%s=`, desktopFileI18nPattern)),
-	regexp.MustCompile(`^Icon=`),
-	regexp.MustCompile(`^Hidden=`),
-	regexp.MustCompile(`^OnlyShowIn=`),
-	regexp.MustCompile(`^NotShowIn=`),
-	regexp.MustCompile(`^Exec=`),
-	regexp.MustCompile(`^TryExec=`),
-	regexp.MustCompile(`^Terminal=`),
-	regexp.MustCompile(`^Actions=`),
-	regexp.MustCompile(`^MimeType=`),
-	regexp.MustCompile(`^Categories=`),
-	regexp.MustCompile(fmt.Sprintf(`^Keywords%s=`, desktopFileI18nPattern)),
-	regexp.MustCompile(`^StartupNotify=`),
-	regexp.MustCompile(`^StartupWMClass`),
-}
-
-func isValidDesktopFilePrefix(line string) bool {
-	for _, re := range validDesktopFileLines {
-		if re.MatchString(line) {
-			return true
-		}
-	}
-	return false
-}
-
-func validExecLine(m *snapYaml, line string) bool {
-	if !strings.HasPrefix(line, "Exec=") && !strings.HasPrefix(line, "TryExec=") {
-		return true
-	}
-	cmd := strings.SplitN(line, "=", 2)[1]
-	for _, app := range m.Apps {
-		validCmd := filepath.Base(generateBinaryName(m, app))
-		// just check the prefix to allow %flag style args
-		// this is ok because desktop files are not run through sh
-		if strings.HasPrefix(cmd, validCmd) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func sanitizeDesktopFile(m *snapYaml, realBaseDir string, rawcontent []byte) []byte {
-	newContent := []string{}
-
-	scanner := bufio.NewScanner(bytes.NewReader(rawcontent))
-	for scanner.Scan() {
-		line := scanner.Text()
-		// ignore everything we have not whitelisted
-		if !isValidDesktopFilePrefix(line) {
-			continue
-		}
-		if !validExecLine(m, line) {
-			continue
-		}
-
-		// do variable substitution
-		line = strings.Replace(line, "${SNAP}", realBaseDir, -1)
-		newContent = append(newContent, line)
-	}
-
-	return []byte(strings.Join(newContent, "\n"))
-}
-
-func addPackageDesktopFiles(m *snapYaml, baseDir string) error {
-	if err := os.MkdirAll(dirs.SnapDesktopFilesDir, 0755); err != nil {
-		return err
-	}
-
-	desktopFiles, err := filepath.Glob(filepath.Join(baseDir, "meta", "gui", "*.desktop"))
-	if err != nil {
-		return fmt.Errorf("cannot get desktop files for %v: %s", baseDir, err)
-	}
-
-	for _, df := range desktopFiles {
-		content, err := ioutil.ReadFile(df)
-		if err != nil {
-			return err
-		}
-
-		realBaseDir := stripGlobalRootDir(baseDir)
-		content = sanitizeDesktopFile(m, realBaseDir, content)
-
-		installedDesktopFileName := filepath.Join(dirs.SnapDesktopFilesDir, fmt.Sprintf("%s_%s", m.Name, filepath.Base(df)))
-		if err := helpers.AtomicWriteFile(installedDesktopFileName, []byte(content), 0755, 0); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func removePackageDesktopFiles(m *snapYaml) error {
-	glob := filepath.Join(dirs.SnapDesktopFilesDir, m.Name+"_*.desktop")
-	activeDesktopFiles, err := filepath.Glob(glob)
-	if err != nil {
-		return fmt.Errorf("cannot get desktop files for %v: %s", glob, err)
-	}
-	for _, f := range activeDesktopFiles {
-		os.Remove(f)
 	}
 
 	return nil
